pyproject.toml
[tool.poetry]
name = "veccache"
version = "0.1.0"
description = ""
authors = ["kaz <kaz.j11@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
openai = "^0.27.8"
faiss-cpu = "^1.7.4"
python-dotenv = "^1.0.0"
numpy = "^1.25.2"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
pytest-mock = "^3.11.1"
pre-commit = "^3.3.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[build]
include = ["vec_cache/**"]
exclude = ["examples/**"]

[tool.isort]
profile = "black"
